+++ src/main/java/org/kakaoshare/backend/domain/product/dto/DetailResponse.java
@@ -6,7 +6,6 @@ import org.kakaoshare.backend.domain.brand.entity.Brand;
 import org.kakaoshare.backend.domain.option.entity.Option;
 import org.kakaoshare.backend.domain.product.entity.Product;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -14,7 +13,7 @@ import java.util.List;
 public class DetailResponse {
     private final Long productId;
     private final String name;
-    private final BigDecimal price;
+    private final Long price;
     private final String type;
     private final String productName;
     private final Boolean hasPhoto;
